Add remove and Add function
